+++ packages/plugin-drizzle/tsconfig.json
@@ -21,5 +21,5 @@
         "moduleDetection": "force",
         "allowArbitraryExtensions": true
     },
-    "include": ["src/**/*.ts", "src/migrations.ts", "src/schema.ts"],
+    "include": ["src/**/*.ts"]
 }
